backend/requirements.txt
alembic==1.12.1
annotated-types==0.7.0
anyio==3.7.1
appdirs==1.4.4
async-timeout==5.0.1
asyncpg==0.29.0
backoff==2.2.1
backtrader==1.9.78.123
bcrypt==3.2.2
beautifulsoup4==4.13.4
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cryptography==45.0.5
cycler==0.12.1
ecdsa==0.19.1
fastapi==0.104.1
fonttools==4.59.0
frozendict==2.4.6
greenlet==3.2.3
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httptools==0.6.4
httpx==0.25.2
idna==3.10
kiwisolver==1.4.8
lxml==6.0.0
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.3
monotonic==1.6
multitasking==0.0.12
numpy==1.26.4
packaging==25.0
pandas==2.1.4
passlib==1.7.4
pillow==11.3.0
posthog==3.1.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.24.0
uvloop==0.21.0
watchfiles==1.1.0
webencodings==0.5.1
websockets==15.0.1

